refactor(admin): extract TeamsTab into its own file
